fleshed out the basic interpreter significantly and started on the c_codegen
I significantly expanded the capibalites of the interpreter
it has a bunch of math functions now and the ability
to work with lists. I added the if expression and
fixed an issue with setting variables multiple times
I did various things internally to accomplish this
but nothing significant worth noting.

Things are really coming along nicely, I'm really exicted