added some basic tests and refactored prim.c and prim.h as stated in last commit
I changed prim.h to just declarations. funtions and function objects are
now in prim.c which is linked into the final executable.
This cleaned up both files, prim.h is just macro definations and
called to the DEFUN macro to declare stuff. and prim.c has fewer
macro definations and more function definations, allong with
a bunch of calls to a different DEFUN macro to make function objects