changed most instances of env to env* and got a very basic llvm function working
I'm hoping changing environments to pointers will help with recursive
functions and such, but I'm not sure. None the less it makes the
interface to environments more consistent.

I couldn't manage to get a version of sum(defun sum (x) (reduce x +)) to
work but I did get a trivial fma function(* x (+ y z)) or so, to work,
which was satisfying, and moreso it means I understand llvm somewhat