added ++ and -- functions, unlike in c they don't modify the argument
I also realized an issue, I can't change variables from functions (which
makes sense as everything is pass by value and all). I see a couple ways
of dealing with this, the more lispy way is to make sure the symbol is
passed as the argument(i.e. require it to be quoted) which is probably
what I'll do(or you know define it as a macro instdead) but that
requires a bit more work on the metaprogramming side before I can add
that. The other way is to use explict referances like in sml, but I
doubt I'll do that.