added nested comment support in lexer, also fixed an issue with prog1
Comments are at the moment entierly eliminated during lexing, whch is
probably fine, I could keep them for context but that seems like a bunch
of unessessary work.

The comments do nest, which is good, I really like nested comments, its
really annoying in C to comment a region and have a close comment in
that region mess stuff up.

As for prog1 I just needed to evaluate the car of a cons cell rather
than the cons cell itself at one point