lots of work on writing generic arathmatic functions, a bit of work on c-codegen and work on simplifying prim generation
Most of this commit is work on writing bignum functions which work with
arbitary arguments(as of now they still require bignum 1st args
however). The ultimate goal of this is to write arathmatic functions
which act on an arbitary number of arguments and return a value of the
type of the first argument(so if the first arg was a double return a
double(i.e. cast any long to a double) or return error if an argument
can't be cast to the type of the first argument.

I tried working on c-codegen again. I really need a more orginized plan
on how to do this, I need to actually use some sort of referance rather
than just winging it, I don't think I can write codegen in the same was
as eval.

I also wrote a bunch of elisp code to help generate prims (well really
to help with helping generate prims)