got a working arathmatic driver, woo!
So as of now arathmatic functions are determined by the type of their
first arg, I'm actually ok with this as it keeps a decent balance of
dynamic types and actualy writable code.

I've named the functions add,sub,mul...etc for now. I need to decide
wether or not I should replace +,-,*...etc with these.