Added gen_cffi which will (hopefully) generate a set of Equivalent SciLisp functions given a C header file.
I had to add clang as an llvm submodule because I need it to parse the C
files, because writing a parser for C is not something I need or want to
do. I haven't finished it yet, I still need to add a bunch of
information that assiocates SciLisp stuff(tags,typechecking
functions,etc) with ctypes. And I still need to add a couple of
functions which I left undefined for now.