Worked on regular expressions, finished writing the code, but didn't have time to test things.
I wrote the regex stuff in such a way that the lisp interface is fairly
high level relitive to what's going on in c. I also changed the actual
emacs_regex.c file to use gc for allocation to make life eaiser.

Also I added a few macros to format type error messages in a standard
easy to used way.