A lot of things, changed default array to a dynamicly typed array, added in a bunch of tree code, changed arrays in all bits of code to use new type of array
Previously the basic array was an array of `union data` values which all
had to be the same type and caused issues with using higher order
functions on arrays. So now the default is an array of sexps and the
previous array type is now refered to as a typed array. I have a working
quick sort for this array type, as well as a working aref, I have
written array->list and list->array functinos but I havent tested them
yet.

I added a couple trivial but important primitive functions, list to
allow evaluating values of a new list and not which is just a boolean
not function.

One syntax note is that typed arrays are, for now double bracketed, so
[[1 2 3]] is an array of longs and [1 2 3] is an array of sexps.

I think I have the basic implementatino of a naieve binary tree done,
but I need to test it.