added a few list functions, updated gc to version 7.4.0
I'm not sure if The gc update is something that will get reflected in
the commit or not.

I added split,drop and take functions on lists. split returns a cons
cell who's cdr is a (new) list of the first n elements of a given list
and who's cdr is the remaining elements. It takes an optional argument
for n, if n is NIL it uses the middle of the list as a split point.
take and drop have the usual take and drop semantics.

I also refined my reverse function, before I only had a (not working)
destructive reverse function, now I have a working destructive reverse
function and a non destructive reverse function.