improved threading in frontend by using pthread_once. Wrote rehash function for hashtable and did a little work towards implementing keyword arguments in the front end.
Using pthread_once not only cleaned up my code a bunch(and eliminated a
mutex and a cond_lock) but also seems to have improved the relilability
of multithreading, I have some idea of why things might be crashing, I
should look into cord io and see if it is multithreaded or not.

I'm pretty much done writing the hash table stuff, now I just need to
test it and integreate it with the rest of the standard library.

I'm going to use an alist for keyword arguments. The last argument will
just be named #keyargs and the value will be the alist of actual keyword
/ variable pairs, I'll set whatever keyword arguments get passed and the
rest will have a default value(either nil or user specified)