Worked on trees added a basic queue implementation in cons.c and added equality comparisons for arrays and conses in lisp_equal
I think I've mostly got basic trees written, I wrote traversal functions
and a map function for trees and somewhat cleaned up my older tree
functions.

The queue thing isn't actually a structure or anything I just made an
fifo queue out of a cons cell. the car is the head of the queue/the list
that composes the queue and the cdr is the last cons cell in the
queue. that way its easy to append an element to the tail of the list
and to pull of the first element, which is kinda cool