exciting update, spellchecking..., yep, I went through a bunch of files and corrected typos and misspellings (mostly in error messages, but in some comments too)
The other stuff I did didn't actually change anything, but I have a
couple ideas of stuff to do. I want to give functions docstrings, so I
added a field for that in the function and macro structures and the
generate-prims alist. I'll start with just the function signature for
documentation because even that can be really useful.

Also I'm probably going to rewrite lexing for things of the form
def<something> (i.e defmacro,defun,defconst,etc..) and I wrote some code
to do that in a comment in lisp.lex, I'll test it out next time