Added autotools to build, also apperently removed gmp from the tree (which is for the best)
So using autotools for the build system was a lot eaiser than I'd
anticipated, which is nice. I still have some hacking to do on building
subprojects (specifically gc and cord, which I want to link into the
main library). I also changed things so the final SciLisp executable is
just frontend.c linked with everything else which is all in a libtool
library libSciLisp.la