removed the remaning functions from primc_header.c and into lisp_types.c, and changed condidionals to use autoconf (specifically DEBUG and MULTITHREADED)
I changed the macros I was using for hacky options (DEBUG,
MULTI_THREADED, GC_THREADS, HAVE_READLINE). and changed them to be
options specified by configure. for gc-theads and readline I use --with
and for everything else uses --enable. Its quite nice. I also changed
the -t and -r (the test options) to take an optional argument which is a
file of tests. I'm going to use automake to run my tests (eventually)
and that'll help wiht that.