Wrote up initial code for format, mostly a copy of CORD_sprintf with code added to acount for lisp stuff.
Specifically I need to change code that used va_args to use a lisp rest
arg which requires for each arg checking we have an arg and then
typechecking that arg. (I suppose I added typechhecking too, as printf
doesn't typecheck it's args). I also removed the %n specifier (because it
writes to memory and is just kinda weird), added an %a argument to print
a lisp object and changed %s to print a lisp object. I'm not sure if
this will work or not...