work on llvm codegen and the new interpreter, also rewrote the functions used to parse escape sequences
I didn't do much work on llvm, but I started writing codegen for atoms.

I wrote code in the interpreter for calling c functions.

I'll need to rewrite c functions that take a rest arg to use a new
format where they take an integer specifing the number of arguments and
a pointer to those arguments.