not much, just some code to convert strings to arrays
I decided to convert a string to an array of characters or wide
characters and not an arary of bytes, for now at least. This will
hopefully make sequence functions make sense on strings(which is the
whole reason for the conversion in the first place). but if not I may
just convert strings into arrays of bytes