More work on the reader, specifically on reading symbols/numbers and lists
I put code to get a symbol from a package into the reader, which in
retrospect was a bad idea, so I need to move it. This means I need to
figure out a way to assoicate symbols with a package. Just adding
another field to the symbol struct might work. I didn't do that before
because it caused environments to not get gced because of circular
references. But I think it should be ok now with how I've re done
symbols and environments.

I wrote read_list. I need to write read_vector and figure out how I'm
ultimately going to handle input and then I should be done (at least the
most basic stuff, I still have some # syntax I want to add.