Some various work on strings and streams, and added a really basic record type
I just did a bunch of fiddling around with streams, trying to write code
for a read-write cord stream, which probably won't work as is.

I added a few string utility functions like stripping/trimming
characters from the start/end of a string, or capatilizing ascii chars.

The record type I added is just a really simple assoicative array, which
maps symbols to sexps. but it should work as a means of having
some object with named fields.

I also wrote a bunch of stuff about the type heirarchy in a comment at
the end of types.h