Some work on read_test, which is starting to look more like a simple repl
I added interpreter.c to the read test sources and got everything to
compile but there are a lot of bugs, which is something I just need to
take some time and work through, since I actually have something to
test.

I don't add any functions to the symbol table yet so I need to do that
as well to get the interpreter working.