Some minor changes to frontend.c to make it compile, also added some new tests
I also did a little work on threads and atomic operations, but nothing
major. A added two tests, a sanity test which has a file with expected
output which should be checked by another program (e.g perl), this is to
insure that the basic mechanics of the language work so that other tests
which rely on the language can be trusted. Also a file which prints out
the emacs hello file line by line to test unicode printing.